# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "whell-laravel-api-documentation",
        "version": "v1.0.0"
    },
    "paths": {
...
        "/api/admin/user/group/{id}": {
            "get": {
                "tags": [
                    "用户组"
                ],
                "summary": "获取用户组",
                "description": "获取用户组",
                "operationId": "App\\Http\\Controllers\\Api\\Admin\\User\\GroupController::get",
                "parameters": [
                    {
                        "name": "accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/accept"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "用户组id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "0": {
                                            "$ref": "#/components/schemas/GroupData"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "用户组"
                ],
                "summary": "删除用户组",
                "description": "删除用户组",
                "operationId": "App\\Http\\Controllers\\Api\\Admin\\User\\GroupController::delete",
                "parameters": [
                    {
                        "name": "accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/accept"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "用户组id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "成功状态 1成功 0失败",
                                    "type": "integer",
                                    "example": 1
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/user/group": {
            "get": {
                "tags": [
                    "用户组"
                ],
                "summary": "获取用户组列表",
                "description": "获取用户组列表",
                "operationId": "App\\Http\\Controllers\\Api\\Admin\\User\\GroupController::search",
                "parameters": [
                    {
                        "name": "accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/accept"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "0": {
                                            "$ref": "#/components/schemas/GroupData"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "用户组"
                ],
                "summary": "更新用户组",
                "description": "更新用户组",
                "operationId": "App\\Http\\Controllers\\Api\\Admin\\User\\GroupController::update",
                "parameters": [
                    {
                        "name": "accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/accept"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "用户组名称",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rules",
                        "in": "query",
                        "description": "用户组规则",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "状态 1启用 0禁用",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "成功状态 1成功 0失败",
                                    "type": "integer",
                                    "example": 1
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "用户组"
                ],
                "summary": "创建用户组",
                "description": "创建用户组",
                "operationId": "App\\Http\\Controllers\\Api\\Admin\\User\\GroupController::create",
                "parameters": [
                    {
                        "name": "accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/accept"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "用户组名称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rules",
                        "in": "query",
                        "description": "用户组规则",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "成功状态 1成功 0失败",
                                    "type": "integer",
                                    "example": 1
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "GroupData": {
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "名称",
                        "type": "string",
                        "example": "group1"
                    },
                    "status": {
                        "description": "状态 1启用 0禁用",
                        "type": "integer",
                        "example": 1
                    },
                    "rules": {
                        "description": "规则",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Index"
                        }
                    }
                },
                "type": "object"
            },
            "accept": {
                "default": "application/whell.api+json"
            },
...
        }
    },
    "security": [
        []
    ]
}